Move prefix range filter from obsolete prefix_range_filter.hpp to new tablefilter_internal_functions
