+++ src/include/duckdb/planner/filter/prefix_range_filter.hpp
@@ -8,35 +8,16 @@
 
 #pragma once
 
-#include "duckdb/common/enums/filter_propagate_result.hpp"
 #include "duckdb/common/optional_ptr.hpp"
 #include "duckdb/common/serializer/deserializer.hpp"
 #include "duckdb/common/serializer/serializer.hpp"
 #include "duckdb/common/types.hpp"
-#include "duckdb/common/uhugeint.hpp"
-#include "duckdb/common/unique_ptr.hpp"
+#include "duckdb/common/types/value.hpp"
 #include "duckdb/planner/table_filter.hpp"
 
 namespace duckdb {
 
-class PrefixRangeFilter {
-public:
-	struct BuildState {
-		virtual ~BuildState() = default;
-	};
-
-	virtual ~PrefixRangeFilter() = default;
-	virtual void Initialize(ClientContext &context, idx_t number_of_rows, Value min, Value max) = 0;
-	virtual unique_ptr<BuildState> InitializeBuildState(ClientContext &context) const = 0;
-	virtual void InsertKeys(Vector &keys, idx_t count, BuildState &state) const = 0;
-	virtual void MergeBuildState(BuildState &state) = 0;
-	virtual idx_t LookupKeys(Vector &keys, SelectionVector &result_sel, idx_t count) const = 0;
-	virtual bool LookupOneValue(const Value &key) const = 0;
-	virtual FilterPropagateResult LookupRange(const Value &lower_bound, const Value &upper_bound) const = 0;
-	virtual bool IsInitialized() const = 0;
-	static unique_ptr<PrefixRangeFilter> CreatePrefixRangeFilter(const LogicalType &key_type);
-	static bool TryComputeSpan(const Value &lower_bound, const Value &upper_bound, uhugeint_t &result);
-};
+class PrefixRangeFilter;
 
 //! DEPRECATED - only preserved for backwards-compatible deserialization and expression conversion
 class PrefixRangeTableFilter final : public TableFilter {
@@ -49,7 +30,6 @@ private:
 public:
 	static constexpr auto TYPE = TableFilterType::PREFIX_RANGE_FILTER;
 	static bool SupportedType(const LogicalType &type);
-	static unique_ptr<PrefixRangeFilter> CreateBitmap(const LogicalType &type);
 
 public:
 	explicit PrefixRangeTableFilter(optional_ptr<PrefixRangeFilter> filter_p, const string &key_column_name_p,
